issue#444 move object definitions to lunatik_obj.h
